feat(UserService): delete exception handler

+++ src/main/java/com/isaiasgabriel/spring_backend/services/UserService.java
@@ -2,8 +2,11 @@ package com.isaiasgabriel.spring_backend.services;
 
 import com.isaiasgabriel.spring_backend.entities.User;
 import com.isaiasgabriel.spring_backend.repositories.UserRepository;
+import com.isaiasgabriel.spring_backend.services.exceptions.DatabaseException;
 import com.isaiasgabriel.spring_backend.services.exceptions.ResourceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -31,7 +34,16 @@ public class UserService {
     }
 
     public void delete(Long id){
-        repository.deleteById(id);
+        try {
+            if (!repository.existsById(id)) {
+                throw new ResourceNotFoundException(id);
+            }
+            repository.deleteById(id);
+        } catch (EmptyResultDataAccessException e) {
+            throw new ResourceNotFoundException(id);
+        } catch (DataIntegrityViolationException e) {
+            throw new DatabaseException(e.getMessage());
+        }
     }
 
     public User update(Long id, User obj) {
